fix build 21.3

+++ src/Types/Decimal.php
@@ -1,14 +1,14 @@
 <?php
 
 /**
- * Class Float.
+ * Class Decimal.
  * 
  * @author Jeremie Litzler
  * @copyright Copyright (c) 2015
  * @licence http://opensource.org/licenses/gpl-license.php GNU Public License
  * @link https://github.com/Puzzlout/Objects
  * @since Version 1.0.0
- * @package Float
+ * @package Decimal
  * @see http://php.net/manual/en/language.types.intro.php
  */
 
@@ -16,7 +16,7 @@ namespace Puzzlout\Objects\Types;
 
 use Puzzlout\Objects\Interfaces\IObject;
 
-class Float extends Integer implements IObject {
+class Decimal extends Integer implements IObject {
 
     public function GetClass() {
         parent::GetClass();
